LPD-91502 Rename DashboardsContext to InventoryContext
