+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/components/AllStructureTypesDropdown.tsx
@@ -9,7 +9,7 @@ import React, {useContext, useEffect, useState} from 'react';
 
 import ApiHelper from '../../../common/services/ApiHelper';
 import getLocalizedValue from '../../../common/utils/getLocalizedValue';
-import {DashboardsContext} from '../DashboardsContext';
+import {InventoryContext} from '../InventoryContext';
 import {Item} from './FilterDropdown';
 import {IAllFiltersDropdown, initialFilters} from './InventoryAnalysisCard';
 import PickerTrigger from './PickerTrigger';
@@ -19,7 +19,7 @@ const AllStructureTypesDropdown: React.FC<IAllFiltersDropdown> = ({
 	item,
 	onSelectItem,
 }) => {
-	const {constants} = useContext(DashboardsContext);
+	const {constants} = useContext(InventoryContext);
 
 	const [structures, setStructures] = useState<Item[]>([
 		initialFilters.structure,
