+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/components/ContentAndFilesCard.tsx
@@ -20,7 +20,7 @@ import {sub} from 'frontend-js-web';
 import React, {useContext, useEffect, useMemo, useState} from 'react';
 
 import ApiHelper from '../../../common/services/ApiHelper';
-import {DashboardsContext} from '../DashboardsContext';
+import {InventoryContext} from '../InventoryContext';
 
 export interface IMetricsProps {
 	categoriesCount: number;
@@ -46,7 +46,7 @@ const ContentAndFilesCard: React.FC<IContentAndFilesCard> = ({
 }) => {
 	const {
 		filters: {language, space},
-	} = useContext(DashboardsContext);
+	} = useContext(InventoryContext);
 
 	const [loading, setLoading] = useState(true);
 	const [metrics, setMetrics] = useState<IMetricsProps>();
