+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/dashboard/components/InventoryAnalysisCard.tsx
@@ -10,7 +10,7 @@ import {buildQueryString} from '@liferay/analytics-reports-js-components-web';
 import React, {useContext, useEffect, useState} from 'react';
 
 import ApiHelper from '../../../common/services/ApiHelper';
-import {DashboardsContext} from '../DashboardsContext';
+import {InventoryContext} from '../InventoryContext';
 import usePagination from '../utils/usePagination';
 import {AllCategoriesDropdown} from './AllCategoriesDropdown';
 import {AllStructureTypesDropdown} from './AllStructureTypesDropdown';
@@ -151,7 +151,7 @@ const dropdownItems: DropdownItem[] = [
 export function InventoryAnalysisCard() {
 	const {
 		filters: {language, space},
-	} = useContext(DashboardsContext);
+	} = useContext(InventoryContext);
 
 	const [filters, setFilters] = useState<{
 		category: Item;
